+++ examples/teensy4_blinky/examples/with_logs.rs
@@ -13,9 +13,8 @@ use bsp::logging;
 
 use embedded_hal::serial::Write;
 
-use rtic_monotonics::imxrt::Gpt1 as Mono;
-use rtic_monotonics::imxrt::*;
-use rtic_monotonics::Monotonic;
+use rtic_monotonics::imxrt::prelude::*;
+imxrt_gpt1_monotonic!(Mono, board::PERCLK_FREQUENCY);
 
 #[rtic::app(device = teensy4_bsp, dispatchers = [LPSPI1])]
 mod app {
@@ -61,8 +60,7 @@ mod app {
 
         // Initialize Monotonic
         gpt1.set_clock_source(hal::gpt::ClockSource::PeripheralClock);
-        let gpt1_mono_token = rtic_monotonics::create_imxrt_gpt1_token!();
-        Mono::start(board::PERCLK_FREQUENCY, gpt1.release(), gpt1_mono_token);
+        Mono::start(gpt1.release());
 
         // Setup LED
         let led = board::led(&mut gpio2, pins.p13);