+++ rtic/CHANGELOG.md
@@ -7,6 +7,9 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ## [Unreleased]
 
+### Changed
+
+- Remove unused dependency `rtic-monotonics`
 ## [v2.1.1] - 2024-03-13
 
 ### Fixed